refactor: use `components` export from Slice Machine
